--- tests/st/bert/utils.py
@@ -1,54 +1,0 @@
-# Copyright 2020 Huawei Technologies Co., Ltd
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ============================================================================
-
-"""
-Functional Cells used in Bert finetune and evaluation.
-"""
-import math
-from tinyms.callbacks import Callback
-
-
-
-class LossCallBack(Callback):
-    """
-    Monitor the loss in training.
-    If the loss in NAN or INF terminating training.
-    Note:
-        if per_print_times is 0 do not print loss.
-    Args:
-        per_print_times (int): Print loss every times. Default: 1.
-    """
-    def __init__(self, dataset_size=-1):
-        super(LossCallBack, self).__init__()
-        self._dataset_size = dataset_size
-    def step_end(self, run_context):
-        """
-        Print loss after each step
-        """
-        cb_params = run_context.original_args()
-        if self._dataset_size > 0:
-            percent, epoch_num = math.modf(cb_params.cur_step_num / self._dataset_size)
-            if percent == 0:
-                percent = 1
-                epoch_num -= 1
-            print("epoch: {}, current epoch percent: {}, step: {}, outputs are {}"
-                  .format(int(epoch_num), "%.3f" % percent, cb_params.cur_step_num, str(cb_params.net_outputs)))
-        else:
-            print("epoch: {}, step: {}, outputs are {}".format(cb_params.cur_epoch_num, cb_params.cur_step_num,
-                                                               str(cb_params.net_outputs)))
-
-
-
-
